fixed the migration file

+++ db/migrate/20160526174229_create_students.rb
@@ -2,8 +2,8 @@ class CreateStudents < ActiveRecord::Migration
   def change
     create_table :students do |t|
       t.string :name
-      t.decimal :weight
-      t.decimal :height
+      t.integer :weight
+      t.integer :height
       t.string :color
       t.decimal :gpa
